web_requirements.txt
# متطلبات الخادم الإضافية لواجهة الويب
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
numpy>=1.21.0
opencv-python>=4.5.0
torch>=1.7.1
torchvision>=0.8.2
basicsr>=1.4.2
facexlib>=0.3.0
realesrgan>=0.3.0
gfpgan>=1.3.8